+++ iroh/examples/rpc.rs
@@ -9,7 +9,7 @@
 use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};
 
 use iroh::collection::IrohCollectionParser;
-use iroh::database::flat2;
+use iroh::database::flat;
 use iroh::rpc_protocol::{ProviderRequest, ProviderResponse};
 use iroh::{bytes::util::runtime, rpc_protocol::ProviderService};
 use iroh_net::tls::Keypair;
@@ -47,7 +47,7 @@ async fn main() -> anyhow::Result<()> {
     let iroh_data_dir = std::env::current_dir()?.join(".iroh");
     tokio::fs::create_dir_all(&iroh_data_dir).await?;
     // create a new persistent database
-    let db = flat2::Database::load(iroh_data_dir.clone(), iroh_data_dir).await?;
+    let db = flat::Database::load(iroh_data_dir.clone(), iroh_data_dir).await?;
     // create a new iroh runtime with 1 worker thread, reusing the existing tokio runtime
     let rt = runtime::Handle::from_currrent(1)?;
     // create a random keypair
