+++ iroh/src/database.rs
@@ -1,5 +1,7 @@
 //! Various database implementations for storing blob data
 #[cfg(feature = "flat-db")]
-pub mod flat2;
+pub mod flat;
 #[cfg(feature = "mem-db")]
 pub mod mem;
+
+pub mod test;
