+++ iroh/tests/provide.rs
@@ -15,7 +15,6 @@ use futures::{
 };
 use iroh::{
     collection::{ArrayLinkStream, Blob, Collection, IrohCollectionParser},
-    database::mem,
     node::{Builder, Event, Node, StaticTokenAuthHandler},
 };
 use iroh_io::{AsyncSliceReader, AsyncSliceReaderExt};
@@ -149,25 +148,23 @@ fn get_options(peer_id: PeerId, addrs: Vec<SocketAddr>) -> iroh::dial::Options {
     }
 }
 
-#[cfg(feature = "legacy-flat-db")]
+#[cfg(feature = "mem-db")]
 #[tokio::test(flavor = "multi_thread")]
 async fn multiple_clients() -> Result<()> {
-    use iroh::database::flat::{create_collection, DataSource};
-
-    let dir: PathBuf = testdir!();
-    let filename = "hello_world";
-    let path = dir.join(filename);
     let content = b"hello world!";
     let addr = "127.0.0.1:0".parse().unwrap();
 
-    tokio::fs::write(&path, content).await?;
-    // hash of the transfer file
-    let expect_data = tokio::fs::read(&path).await?;
-    let expect_hash = blake3::hash(&expect_data);
-    let expect_name = filename.to_string();
-
-    let (db, hash) = create_collection(vec![DataSource::new(path)]).await?;
-
+    let mut db = iroh::database::test::Database::default();
+    let expect_hash = db.insert(content.as_slice());
+    let expect_name = "hello_world".to_string();
+    let collection = Collection::new(
+        vec![Blob {
+            name: expect_name.clone(),
+            hash: expect_hash,
+        }],
+        0,
+    )?;
+    let hash = db.insert(collection.to_bytes()?);
     let rt = test_runtime();
     let node = test_node(db, addr).runtime(&rt).spawn().await?;
 
@@ -229,7 +226,7 @@ where
     let mut expects = Vec::new();
     let num_blobs = file_opts.len();
 
-    let (mut mdb, lookup) = mem::Database::new(file_opts.clone());
+    let (mut mdb, lookup) = iroh::database::test::Database::new(file_opts.clone());
     let mut blobs = Vec::new();
     let mut total_blobs_size = 0u64;
 
@@ -364,7 +361,7 @@ async fn test_server_close() {
     let rt = test_runtime();
     // Prepare a Provider transferring a file.
     setup_logging();
-    let mut db = iroh::database::mem::Database::default();
+    let mut db = iroh::database::test::Database::default();
     let child_hash = db.insert(b"hello there");
     let collection = Collection::new(
         vec![Blob {
@@ -468,8 +465,8 @@ async fn test_blob_reader_partial() -> Result<()> {
 /// returns the database and the root hash of the collection
 fn create_test_db(
     entries: impl IntoIterator<Item = (impl Into<String>, impl AsRef<[u8]>)>,
-) -> (mem::Database, Hash) {
-    let (mut db, hashes) = iroh::database::mem::Database::new(entries);
+) -> (iroh::database::test::Database, Hash) {
+    let (mut db, hashes) = iroh::database::test::Database::new(entries);
     let collection = Collection::new(
         hashes
             .into_iter()
@@ -668,7 +665,7 @@ async fn test_custom_collection_parser() {
     // create a collection consisting of 2 leafs
     let leaf1_data = vec![0u8; 12345];
     let leaf2_data = vec![1u8; 67890];
-    let mut db = mem::Database::default();
+    let mut db = iroh::database::test::Database::default();
     let leaf1_hash = db.insert(leaf1_data.clone());
     let leaf2_hash = db.insert(leaf2_data.clone());
     let collection = vec![leaf1_hash, leaf2_hash];
@@ -741,7 +738,7 @@ impl CustomGetHandler for BlobCustomHandler {
 async fn test_custom_request_blob() {
     let rt = test_runtime();
     let expected = b"hello".to_vec();
-    let (db, hashes) = iroh::database::mem::Database::new([("test", &expected)]);
+    let (db, hashes) = iroh::database::test::Database::new([("test", &expected)]);
     let hash = Hash::from(*hashes.values().next().unwrap());
     let addr = "127.0.0.1:0".parse().unwrap();
     let node = test_node(db, addr)
